code cleanup

+++ SBJEV3/SBJEV3Chunk.h
@@ -34,7 +34,7 @@ public:
 		}
 	}
 	
-	const uint64_t size() const
+	const size_t size() const
 	{
 		return _size;
 	}
